+++ service/internal/eventstreamtesting/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.7.3 (2026-09-04)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.7.2 (2026-08-31.2)
 
 * **Dependency Update**: Updated to the latest SDK module versions